requirements.txt
asgiref==3.8.1
dj-database-url==2.1.0
Django==5.0.6
djangorestframework==3.15.1
iniconfig==2.0.0
packaging==24.0
pluggy==1.5.0
pytest==8.2.0
pytest-django==4.8.0
python-decouple==3.8
sqlparse==0.5.0
typing_extensions==4.11.0
